ImageFactory refactor
