+++ _pylibczi/Image.cpp
@@ -37,7 +37,7 @@ namespace pylibczi {
       std::vector<std::pair<char, int> > ans;
       for (auto di : sort_order) {
           int value;
-          if (m_cdims.TryGetPosition(di, &value)) ans.emplace_back(libCZI::Utils::DimensionToChar(di), value);
+          if (m_planeCoordinates.TryGetPosition(di, &value)) ans.emplace_back(libCZI::Utils::DimensionToChar(di), value);
       }
       if (isMosaic) ans.emplace_back('M', m_mIndex);
       return ans;
@@ -49,7 +49,7 @@ namespace pylibczi {
       std::vector<CZI_DI> sort_order{CZI_DI::S, CZI_DI::T, CZI_DI::C, CZI_DI::Z};
       for (auto di : sort_order) {
           int di_value, other_value;
-          if (m_cdims.TryGetPosition(di, &di_value) && other.m_cdims.TryGetPosition(di, &other_value) && di_value!=other_value)
+          if (m_planeCoordinates.TryGetPosition(di, &di_value) && other.m_planeCoordinates.TryGetPosition(di, &other_value) && di_value!=other_value)
               return (di_value<other_value);
       }
       return m_mIndex<other.m_mIndex;
@@ -57,15 +57,15 @@ namespace pylibczi {
 
   size_t ImageBC::calculate_idx(const std::vector<size_t>& idxs)
   {
-      if (idxs.size()!=m_matrixSizes.size())
-          throw ImageAccessUnderspecifiedException(idxs.size(), m_matrixSizes.size(), "Sizes must match");
+      if (idxs.size()!=m_shape.size())
+          throw ImageAccessUnderspecifiedException(idxs.size(), m_shape.size(), "Sizes must match");
       size_t running_product = 1;
       std::vector<size_t> weights(1, 1);
-      std::for_each(m_matrixSizes.rbegin(), m_matrixSizes.rend()-1, [&weights, &running_product](const size_t len) {
+      std::for_each(m_shape.rbegin(), m_shape.rend()-1, [&weights, &running_product](const size_t len) {
           running_product *= len;
           weights.emplace_back(running_product);
       });
-      std::vector<size_t> prod(m_matrixSizes.size(), 0);
+      std::vector<size_t> prod(m_shape.size(), 0);
       std::transform(idxs.begin(), idxs.end(), weights.begin(), prod.begin(), [](size_t a, size_t b) -> size_t {
           return a*b;
       });
@@ -73,78 +73,5 @@ namespace pylibczi {
       return idx;
   }
 
-  ImageFactory::ConstrMap ImageFactory::s_pixelToImage{
-          {PT::Gray8, [](V_ST shp, PT pt, LCD* dims, IR ir, int m) {
-              return std::shared_ptr<Image<uint8_t> >(new Image<uint8_t>(std::move(shp), pt, dims, ir, m));
-          }},
-          {PT::Bgr24, [](V_ST shp, PT pt, LCD* dims, IR ir, int m) {
-              return std::shared_ptr<Image<uint8_t> >(new Image<uint8_t>(std::move(shp), pt, dims, ir, m));
-          }},
-          {PT::Gray16, [](V_ST shp, PT pt, LCD* dims, IR ir, int m) {
-              return std::shared_ptr<Image<uint16_t> >(new Image<uint16_t>(std::move(shp), pt, dims, ir, m));
-          }},
-          {PT::Bgr48, [](V_ST shp, PT pt, LCD* dims, IR ir, int m) {
-              return std::shared_ptr<Image<uint16_t> >(new Image<uint16_t>(std::move(shp), pt, dims, ir, m));
-          }},
-          {PT::Gray32Float, [](V_ST shp, PT pt, LCD* dims, IR ir, int m) {
-              return std::shared_ptr<Image<float> >(new Image<float>(std::move(shp), pt, dims, ir, m));
-          }},
-          {PT::Bgr96Float, [](V_ST shp, PT pt, LCD* dims, IR ir, int m) {
-              return std::shared_ptr<Image<float> >(new Image<float>(std::move(shp), pt, dims, ir, m));
-          }}
-  };
-
-  size_t
-  ImageFactory::size_of_pixel_type(PT pixel_type)
-  {
-      switch (pixel_type) {
-      case PT::Gray8:
-      case PT::Bgr24:return sizeof(uint8_t);
-      case PT::Gray16:
-      case PT::Bgr48:return sizeof(uint16_t);
-      case PT::Gray32Float:
-      case PT::Bgr96Float:return sizeof(float);
-      default:throw PixelTypeException(pixel_type, "Pixel Type unsupported by libCZI.");
-      }
-  }
-
-  size_t
-  ImageFactory::n_of_channels(libCZI::PixelType pixel_type)
-  {
-      using PT = libCZI::PixelType;
-      switch (pixel_type) {
-      case PT::Gray8:
-      case PT::Gray16:
-      case PT::Gray32Float:return 1;
-      case PT::Bgr24:
-      case PT::Bgr48:
-      case PT::Bgr96Float:return 3;
-      default:throw PixelTypeException(pixel_type, "Pixel Type unsupported by libCZI.");
-      }
-
-  };
-
-  std::shared_ptr<ImageBC> ImageFactory::construct_image(const std::shared_ptr<libCZI::IBitmapData>& pBitmap,
-          const libCZI::CDimCoordinate* cdims,
-          libCZI::IntRect box,
-          int mIndex)
-  {
-      libCZI::IntSize size = pBitmap->GetSize();
-      libCZI::PixelType pt = pBitmap->GetPixelType();
-
-      std::vector<size_t> shp;
-      size_t channels = n_of_channels(pt);
-      if (channels==3)
-          shp.emplace_back(3);
-      shp.emplace_back(size.h);
-      shp.emplace_back(size.w);
-
-      std::shared_ptr<ImageBC> img = s_pixelToImage[pt](shp, pt, cdims, box, mIndex);
-      if (img==nullptr)
-          throw std::bad_alloc();
-      img->load_image(pBitmap, channels);
-
-      return std::shared_ptr<ImageBC>(img);
-  }
 
 }
