+++ _pylibczi/pb_helpers.cpp
@@ -11,6 +11,7 @@
 #include <set>
 #include <iostream>
 #include "Image.h"
+#include "ImageFactory.h"
 #include "Reader.h"
 #include "exceptions.h"
